Add namespaces to tests
