+++ tests/PaymentMathTest.php
@@ -1,7 +1,9 @@
 <?php
 
-use SilverStripe\Omnipay\PaymentMath;
+namespace SilverStripe\Omnipay\Tests;
 
+use SilverStripe\Omnipay\PaymentMath;
+use SilverStripe\Dev\SapphireTest;
 /**
  * Test payment math operations
  */
