+++ tests/PaymentModelTest.php
@@ -1,4 +1,7 @@
 <?php
+
+namespace SilverStripe\Omnipay\Tests;
+
 use SilverStripe\Omnipay\GatewayInfo;
 
 class PaymentModelTest extends PaymentTest
