+++ tests/PaymentTest.php
@@ -1,10 +1,14 @@
 <?php
+
+namespace SilverStripe\Omnipay\Tests;
+
 use SilverStripe\Omnipay\Service\PaymentService;
 use SilverStripe\Omnipay\Service\ServiceFactory;
+use SilverStripe\Dev\FunctionalTest;
 
 abstract class PaymentTest extends FunctionalTest
 {
-    public static $fixture_file = array(
+    protected static $fixture_file = array(
         'payment.yml'
     );
